:fire: remove: 불필요 코드 및 파일삭제
