+++ srcs/Server/includes/Thread.hpp
@@ -3,7 +3,6 @@
 
 #include <iostream>
 #include <thread>
-#include <atomic>
 
 #define MAX_THREAD 2
 
@@ -15,16 +14,17 @@ private:
 
 public:
     template <class _Fp, class... _Args>
-    static void createThread(_Fp&& __f, _Args&&... __args)
+    static int createThread(_Fp&& __f, _Args&&... __args)
     {
         for (int i = 0; i < MAX_THREAD; i++)
         {
             if (thread_arr[i] == nullptr)
             {
                 thread_arr[i] = new std::thread(__f, __args...);
-                break;
+                return EXIT_SUCCESS;
             }
         }
+        return EXIT_FAILURE;
     }
     static void cleanThread();
 };
