+++ srcs/Server/srcs/Thread.cpp
@@ -1,10 +1,10 @@
 #include "Thread.hpp"
 
-// extern std::thread *thread_arr[MAX_THREAD];
-std::thread* Thread::thread_arr[MAX_THREAD] = {nullptr};
+std::thread *Thread::thread_arr[MAX_THREAD] = {nullptr};
 
 void Thread::cleanThread()
 {
+    std::cout << "cleanThread start..." << std::endl;
     for (int i = 0; i < MAX_THREAD; i++)
     {
         if (thread_arr[i] == nullptr)
