+++ srcs/Server/srcs/Util.cpp
@@ -5,6 +5,20 @@
 
 extern std::atomic<bool> atom_stop;
 
+void error_msg(enum e_error flag)
+{
+    if (errno)
+    {
+        std::cerr << "Error: " << std::strerror(errno) << std::endl;
+        exit(errno);
+    }
+    else if (flag == WRONG_ARG)
+        std::cerr << "Error: " << "Wrong number of arguments" << std::endl;
+    else if (flag == FATAL_ERR)
+        std::cerr << "Error: " << "Fatal error" << std::endl;
+    exit(EXIT_FAILURE);
+}
+
 void cleanMemory()
 {
     Thread::cleanThread();
@@ -28,8 +42,6 @@ void signalHandler(int signum) {
     }
     else
         std::cout << "Others: Signal " << signum << " received" << std::endl;
-    atom_stop = true;
-    std::cout << "signalHandler atom_stop = true" << std::endl;
     cleanMemory();
     exit(signum);
 }
